fem route refactor with backward compatiibty
