+++ jno/utils/solver/__init__.py
@@ -2,15 +2,26 @@
 Solver utility layer for jNO backend routing.
 
 This package contains shared helpers used by FEM/FEAX, time routing,
-and solver adapters. It should not contain frontend DSL code.
+and solver adapters.
 """
 
 from .solver_helper import (
+    iter_children,
+    iter_placeholder_children,
     contains_node_type,
     contains_testfunction,
     contains_trialfunction,
+    contains_model_call,
+    contains_model_eval,
+    depends_on_domain_variables,
+    contains_subexpr,
+    unique_by_id,
     sum_terms,
     apply_sign,
+    is_temporal_var,
+    max_temporal_derivative_order,
+    collect_temporal_tags,
+    contains_temporal_derivative,
 )
 
 from .weak_form_helpers import (
@@ -18,10 +29,21 @@ from .weak_form_helpers import (
 )
 
 __all__ = [
+    "iter_children",
+    "iter_placeholder_children",
     "contains_node_type",
     "contains_testfunction",
     "contains_trialfunction",
+    "contains_model_call",
+    "contains_model_eval",
+    "depends_on_domain_variables",
+    "contains_subexpr",
+    "unique_by_id",
     "sum_terms",
     "apply_sign",
     "split_weak_additive_terms",
+    "is_temporal_var",
+    "max_temporal_derivative_order",
+    "collect_temporal_tags",
+    "contains_temporal_derivative",
 ]
